Remove yarn_mirror
I'm pretty confident that actually checking in the mirror is good in
many contexts, but not for this project.

Libraries should not commit yarn.lock files when they depend on semver
ranges like `dep@^1.0.0`, because developers and CI should always test
against the latest packages consumers will use when building and
testing.

Also, checking in the mirror in git-lfs is not a great idea, the files
are small and it takes an exceptionally long time to checkout the
project from git. If you choose to commit the mirror, make sure to not
fall into the same trap i did.

But for now, I will continue to use the yarn_mirror approach for
non-library projects at work. Because it provides a stricter hermicity
promise than the yarn.lock alone.

If you read this commit message, congrats.